Java task

+++ SimpleJavaProject/src/main/java/arina_tolchenitsyna/Main.java
@@ -8,6 +8,6 @@ public class Main {
         System.out.println(a + b);
         System.out.println(a - b);
         System.out.println(a * b);
-        System.out.println((float) a / b);
+        System.out.println((double) a / b);
     }
 }
